# feat: implement Issue #32 - Side Effects Foundation with Performance Optimizations